requirements.txt
attrs==25.3.0
aws-cdk-lib==2.211.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.5.0
boto3==1.40.16
boto3-stubs==1.40.18
botocore==1.40.16
botocore-stubs==1.38.46
cattrs==25.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
constructs==10.4.2
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.113.0
publication==0.0.3
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
typeguard==2.13.3
types-awscrt==0.27.6
types-requests==2.32.4.20250809
types-s3transfer==0.13.0
typing_extensions==4.14.1
urllib3==2.5.0
